fix: sua loi nop bai va kieu du lieu score_percent
- QuizPage: huy debounce truoc khi set isSubmittedRef de tranh race condition
- QuizPage: reset isSubmittedRef khi submit that bai, cho phep thu lai
- QuizPage: chuyen setShowSubmitModal(false) vao trong try block truoc navigate
- Migration 002: doi score_percent tu INTEGER sang NUMERIC(5,1) de luu diem thap phan
- Migration 001: cap nhat khai bao column score_percent

+++ supabase/migrations/001_quiz_sessions.sql
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS private.quiz_sessions (
     correct_count INTEGER NOT NULL DEFAULT 0 CHECK (correct_count >= 0),
     wrong_count INTEGER NOT NULL DEFAULT 0 CHECK (wrong_count >= 0),
     unanswered_count INTEGER NOT NULL DEFAULT 0 CHECK (unanswered_count >= 0),
-    score_percent INTEGER NOT NULL DEFAULT 0 CHECK (score_percent >= 0 AND score_percent <= 100),
+    score_percent NUMERIC(5,1) NOT NULL DEFAULT 0 CHECK (score_percent >= 0 AND score_percent <= 100),
     duration_seconds INTEGER NOT NULL DEFAULT 0 CHECK (duration_seconds >= 0),
     revision INTEGER NOT NULL DEFAULT 1 CHECK (revision >= 1),
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
